rename config and defs with jet_ prefix
